creation of logic update user email and components

+++ staff/florencia-mussis/api/logic/unregisterUser.js
@@ -8,14 +8,14 @@ function unregisterUser(userId, password) {
     /*2. check if user passed password  3. if not, then error*/
     return users.findOne({ _id: new ObjectId(userId) })
         .then(user => {
+            if (!user) throw new Error (`user with id ${userId} not found`)
+
             if (user.password !== password) throw new Error('wrong credentials')
 
             return stickies.deleteMany({ "user": userId })
                 /* 4. if yes, then delete user from db and its stickies*/
                 .then(() => {
-                    return users.deleteOne({
-                        _id: new ObjectId(userId)
-                    })
+                    return users.deleteOne({_id: new ObjectId(userId)})
                 })
         })
 }
